LayoutTests/imported/w3c/web-platform-tests/storage/opaque-origin.https.window-expected.txt


PASS navigator.storage.persisted() in non-sandboxed iframe should not reject
PASS navigator.storage.persisted() in sandboxed iframe should reject with TypeError
FAIL navigator.storage.estimate() in non-sandboxed iframe should not reject assert_equals: navigator.storage.estimate() should not reject expected "no rejection" but got "API access threw"
FAIL navigator.storage.estimate() in sandboxed iframe should reject with TypeError assert_equals: navigator.storage.estimate() should reject with TypeError expected "correct rejection" but got "API access threw"
PASS navigator.storage.persist() in non-sandboxed iframe should not reject
PASS navigator.storage.persist() in sandboxed iframe should reject with TypeError
